Fix merge conflicts

+++ src/game_engine/board/basic.rs
@@ -16,7 +16,7 @@ pub struct BasicBoard {
     pub highlighted: Vec<Location>,
     pub castling_rights : [bool; 4],
     pub en_passant : i8,
-    pub material_score : i8,
+    pub material_score : i32,
 }
 
 impl BasicBoard {
@@ -275,7 +275,7 @@ impl Board for BasicBoard {
         self.en_passant
     }
 
-    fn get_material_score(&self) -> i8 {
+    fn get_material_score(&self) -> i32 {
         self.material_score
     }
 }
